match: 3 libultra-shaped funcs rebuilt (FD0/B4E0/settimer)
Rebuilds of false-match commits ab83001, 90ccfd7, fbbc285. The yaml
flips alone unlocked B4E0 and settimer (both matched first try at
-O1, which the existing ultra_os_*.c routing already provided).

func_80138FD0 needed an additional source-side fix: arg2 was declared
s16 in the .c, but the original used s32 (visible because IDO emitted
`lh $t0, 0x32(sp)` byte-narrow vs base's `lw $t0, 0x30(sp)` —
the original C took an s32 and let the `sh` store truncate, instead
of the .c's pre-truncated s16 spill-and-reload pattern). Switching to
s32 picks up the full-word load.

+++ src/ultra_os_func_80138FD0.c
@@ -15,7 +15,7 @@ typedef struct Foo_s {
 
 extern Foo *D_80042EB4;
 
-void func_80138FD0(s32 arg0, s32 arg1, s16 arg2) {
+void func_80138FD0(s32 arg0, s32 arg1, s32 arg2) {
     register u32 saveMask = func_80038D70();
     D_80042EB4->_w10 = arg0;
     D_80042EB4->_w14 = arg1;
